Split up TopUpModal.tsx
