+++ src/Pages/Onboarding/steps/TopUpModal/TopUpModal.module.scss
@@ -1,5 +1,5 @@
-@import "../../../assets/styles/base/variables";
-@import '../../../assets/styles/base/mixins/media';
+@import "../../../../assets/styles/base/variables";
+@import '../../../../assets/styles/base/mixins/media';
 
 .modal {
   display: flex;
